dailyhoroscope/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-ses
django-environ
gunicorn
python-decouple
psycopg2-binary
pytest
pytest-django
